+++ src/server/api/routes/channel-members.router.js
@@ -1,0 +1,59 @@
+const express = require('express');
+
+const router = express.Router({ mergeParams: true });
+
+// controllers
+
+const channelMembersController = require('../controllers/channel-members.controller');
+
+/**
+ * @swagger
+ * /channel_members:
+ *  post:
+ *    summary: Create a module
+ *    description:
+ *      Will create a module.
+ *    produces: application/json
+ *    parameters:
+ *      - in: body
+ *        name: module
+ *        description: The module to create.
+ *        schema:
+ *          type: object
+ *          required:
+ *            - title
+ *            - startDate
+ *            - endDate
+ *             - classId
+ *          properties:
+ *            title:
+ *              type: string
+ *            startDate:
+ *              type: string
+ *              format: date-time
+ *            endDate:
+ *              type: string
+ *              format: date-time
+ *            classId:
+ *              type: string
+ *    responses:
+ *      201:
+ *        description: Module created
+ *      5XX:
+ *        description: Unexpected error.
+ */
+router.post('/', (req, res) => {
+  channelMembersController
+    .createChannelMember(req.body)
+    .then((result) => res.json(result))
+    .catch((error) => {
+      console.log(error);
+
+      res
+        .status(400)
+        .send('Bad request')
+        .end();
+    });
+});
+
+module.exports = router;
